+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - logrotate: add minsize 100k [PR #2746]
+- cmake: fix alphabetic requirements when some tests are disabled [PR #2766]
 
 ### Changed
 - cmake: fix add_alphabetic_requirements() no-op sort [PR #2758]
@@ -2340,4 +2341,5 @@ If you want to migrate from your manually configured disk autochanger to simply 
 [PR #2742]: https://github.com/bareos/bareos/pull/2742
 [PR #2746]: https://github.com/bareos/bareos/pull/2746
 [PR #2758]: https://github.com/bareos/bareos/pull/2758
+[PR #2766]: https://github.com/bareos/bareos/pull/2766
 [unreleased]: https://github.com/bareos/bareos/tree/master